+++ junos/provider.go
@@ -114,6 +114,7 @@ func Provider() terraform.ResourceProvider {
 			"junos_security_utm_policy":                                 resourceSecurityUtmPolicy(),
 			"junos_security_utm_custom_url_pattern":                     resourceSecurityUtmCustomURLPattern(),
 			"junos_security_utm_profile_web_filtering_juniper_enhanced": resourceSecurityUtmProfileWebFilteringEnhanced(),
+			"junos_security_utm_profile_web_filtering_juniper_local":    resourceSecurityUtmProfileWebFilteringLocal(),
 			"junos_security_zone":                                       resourceSecurityZone(),
 			"junos_static_route":                                        resourceStaticRoute(),
 			"junos_vlan":                                                resourceVlan(),