+++ CHANGELOG.md
@@ -3,6 +3,7 @@ ENHANCEMENTS:
 * add resource junos_security_utm_custom_url_pattern
 * add resource junos_security_utm_policy
 * add resource junos_security_utm_profile_web_filtering_juniper_enhanced
+* add resource junos_security_utm_profile_web_filtering_juniper_local
 
 BUG FIXES:
 * [workflows] fix compile freebsd/arm64 on release